LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.2_White_Space/S7.2_A5_T2-expected.txt

CONSOLE MESSAGE: line 76: SyntaxError: Unrecognized token '\u000B'
S7.2_A5_T2

PASS Expected parsing failure

TEST COMPLETE
